+++ invenio_requests/__init__.py
@@ -20,7 +20,7 @@ from .proxies import (
     current_requests_service,
 )
 
-__version__ = "6.0.0"
+__version__ = "6.1.0"
 
 __all__ = (
     "__version__",