+++ invenio_requests/assets/semantic-ui/js/invenio_requests/search/RequestsSearchLayout.js
@@ -16,13 +16,16 @@ import { GridResponsiveSidebarColumn } from "react-invenio-forms";
 import { SearchBar } from "react-searchkit";
 import { Button, Container, Grid } from "semantic-ui-react";
 
+import { SharedOrMineFilter } from "@js/invenio_requests/components/SharedOrMineFilter";
+
 export const RequestsSearchLayout = ({ config, appName }) => {
   const [sidebarVisible, setSidebarVisible] = React.useState(false);
   return (
     <Container>
       <Grid>
         <Grid.Row>
-          <Grid.Column only="mobile tablet" mobile={3} tablet={1}>
+          <Grid.Column only="computer" computer={4} />
+          <Grid.Column only="mobile tablet" mobile={2} tablet={1}>
             <Button
               basic
               size="medium"
@@ -34,16 +37,26 @@ export const RequestsSearchLayout = ({ config, appName }) => {
           </Grid.Column>
 
           <Grid.Column
-            mobile={13}
+            mobile={14}
+            tablet={7}
+            computer={4}
+            floated="right"
+            className="text-align-right-mobile"
+          >
+            <SharedOrMineFilter />
+          </Grid.Column>
+
+          <Grid.Column
+            mobile={16}
             tablet={4}
             computer={3}
             floated="right"
             className="text-align-right-mobile"
           >
-            <RequestStatusFilter className="rel-mb-1" />
+            <RequestStatusFilter className="rel-mb-1" keepFiltersOnUpdate />
           </Grid.Column>
 
-          <Grid.Column mobile={16} tablet={11} computer={9}>
+          <Grid.Column mobile={16} tablet={16} computer={5}>
             <SearchBar placeholder={i18next.t("Search in my requests...")} />
           </Grid.Column>
         </Grid.Row>