+++ invenio_requests/services/requests/params.py
@@ -9,10 +9,12 @@
 """Search parameter interpreters for requests."""
 from functools import partial
 
+from invenio_records_resources.references import EntityGrant
 from invenio_records_resources.services.records.params import (
     FilterParam,
     ParamInterpreter,
 )
+from invenio_search.engine import dsl
 
 from ...resolvers.registry import ResolverRegistry
 
@@ -78,4 +80,56 @@ class IsOpenParam(ParamInterpreter):
             search = search.filter("term", **{self.field_name: True})
         elif params.get("is_open") is False:
             search = search.filter("term", **{self.field_name: False})
+        return search
+
+
+class SharedOrMyRequestsParam(ParamInterpreter):
+    """Evaluates the 'shared_with_me' parameter for requests."""
+
+    def _generate_my_requests_query(self, identity):
+        """Generate the query for my requests.
+
+        The query will return requests created by the user or requests where the user is the receiver.
+        """
+        created_by = "created_by.user"
+        receiver = "receiver.user"
+        return dsl.Q(
+            "bool",
+            should=[
+                dsl.Q("term", **{created_by: identity.id}),
+                dsl.Q("term", **{receiver: identity.id}),
+            ],
+        )
+
+    def _generate_shared_with_me_query(self, identity):
+        """Generate the query for shared_with_me.
+
+        The query will return requests shared with the user via the topic grants (only via user or group).
+        """
+        allowed_need_methods = {"id", "role"}
+        grants = [
+            EntityGrant("topic", need).token
+            for need in identity.provides
+            if need.method in allowed_need_methods
+        ]
+        my_requests_query = self._generate_my_requests_query(identity)
+        # Topic grants include requests created by the user or the user is the receiver,
+        # so we need to exclude them
+        return dsl.Q("terms", **{"grants": grants}) & ~my_requests_query
+
+    def apply(self, identity, search, params):
+        """Evaluate the shared_with_me parameter on the search.
+
+        If shared_with_me is True, the search will return requests shared with the user via the topic grants (only via user or group).
+        If shared_with_me is False, the search will return requests created by the user or requests where the user is the receiver.
+        """
+        if params.get("shared_with_me") is True:
+            # Shared with me
+            shared_with_me_query = self._generate_shared_with_me_query(identity)
+            search = search.filter(shared_with_me_query)
+        else:
+            # My requests
+            my_requests_query = self._generate_my_requests_query(identity)
+            search = search.filter(my_requests_query)
+
         return search