+++ CHANGES.rst
@@ -9,6 +9,11 @@
 Changes
 =======
 
+Version v6.1.0 (released 2025-03-11)
+
+- feature: add Topic generator for request types. That enables granting
+  permissions to users based on the topic of the request.
+
 Version v6.0.0 (released 2025-02-13)
 
 - Promote to stable release.